+++ pontoon/translate/views.py
@@ -9,7 +9,7 @@ from django.views.decorators.csrf import (
     ensure_csrf_cookie,
 )
 
-from pontoon.base.utils import get_locale_or_redirect, get_project_or_redirect
+from pontoon.base.services import get_locale_or_redirect, get_project_or_redirect
 
 
 @csrf_exempt
